hw4/requirements.txt
catboost==1.2.8
numpy==2.3.4
pandas==2.3.3
scikit-learn==1.7.2
pyarrow==21.0.0
seaborn==0.13.2
matplotlib==3.10.7
kafka-python-ng==2.2.3
psycopg2-binary==2.9.10
streamlit==1.40.2
six==1.16.0
